--- lib/detailsScreen.dart
@@ -1,40 +1,0 @@
-import 'dart:convert';
-import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
-
-import 'main.dart';
-
-class detailsScreen extends StatefulWidget {
-  final Pokemon pokemon;
-
-  const detailsScreen({Key? key, required this.pokemon}) : super(key: key);
-
-  @override
-  State<detailsScreen> createState() => _detailsScreenState();
-}
-
-class _detailsScreenState extends State<detailsScreen> {
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.pokemon.name),
-        centerTitle: true,
-        backgroundColor: Colors.red,
-      ),
-      body: Center(
-        child: Container(
-            height: 300,
-            child: Image.network(
-                'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/' + widget.pokemon.url.toString().split('/')[6] + '.png'
-            )
-        ),
-      ),
-    );
-  }
-}
