--- lib/models/Pokemon.dart
@@ -1,16 +1,0 @@
-class Pokemon {
-  final String name;
-  final String url;
-
-  const Pokemon({
-    required this.name,
-    required this.url,
-  });
-
-  factory Pokemon.fromJson(Map<String, dynamic> json) {
-    return Pokemon(
-      name: json['name'],
-      url: json['url'],
-    );
-  }
-}
